Forward propogation complete
Network can propogate a sample through the network, now need to add
backpropogation as well. Could stand to be broken up into some more
smaller functions, backpropogation could make the Network.go file a bit
ugly. Makes sense that all of that logic would be in that file, though,
so doubtful.

Liking the client interface, may want to make it not quite so "do it all
at once," but for now just want to get the thing working.