Add ReLU and Tanh activation functions
Add application and derivatives for these two activation functions.
Pulled definitions off of Wikipedia, but as far as I can tell they're
accurate. Should function the exact same way that Sigmoid does now.

May consider switching from this structure model to just accepting a
function that takes a function and uses that instead. But I like how
this format correlates both of the functions together, and how the
application and derivative will be related to each other in the
documentation clearly.