Calculate error at input layer, WAY better results
Like the idiot that I am, I was not calculating the error in the input
layer because I had a bound index wrong (> instead of >=). Fixing this
is giving FAR better results, as we're, you know, actually training all
layers of the network.

Also cleaned up a few lines that were completely stupid.