Add customizable Criterion
MSE is used by default, but any structure that adheres to the interface
outlined in Criterion.go can be used to calculate the Error at each
iteration, along with the error in the output layer. May need to rename
things going forward, beacuse to be honest these are pretty names at the
moment. Borrowed some naming from the Torch Lua framework, and
incidentally I cannot spell Criterion correctly the first time for the
life of me.